requirements.txt
Flask
firebase-admin
google-generativeai
python-dotenv
torch
transformers
sentencepiece
numpy
wordcloud
pythainlp
matplotlib
requests
gunicorn